+++ lib/gitflash/git/command_error.rb
@@ -2,6 +2,10 @@
 
 module Gitflash
   module Git
-    class CommandError < StandardError; end
+    class CommandError < Gitflash::Error
+      def initialize(message = nil, code: 'git_failed', plan: nil)
+        super
+      end
+    end
   end
 end